integration: move test_counter_batch() to batch.rs
The test is extracted into `statements/batch.rs`.
